+++ src/dataset/infrastructure/repositories/DatasetJSDataverseRepository.ts
@@ -39,7 +39,6 @@ import {
   deleteDatasetDraft,
   getDatasetCitationInOtherFormats,
   getDatasetAvailableCategories,
-  getDatasetTemplates,
   linkDataset,
   unlinkDataset,
   getDatasetLinkedCollections,
@@ -62,7 +61,6 @@ import { axiosInstance } from '@/axiosInstance'
 import { requireAppConfig } from '../../../config'
 import { AxiosResponse } from 'axios'
 import { JSDataverseReadErrorHandler } from '@/shared/helpers/JSDataverseReadErrorHandler'
-import { DatasetTemplate } from '@/dataset/domain/models/DatasetTemplate'
 import { CollectionSummary } from '@/collection/domain/models/CollectionSummary'
 
 const includeDeaccessioned = true
@@ -417,10 +415,6 @@ export class DatasetJSDataverseRepository implements DatasetRepository {
   }
   getAvailableCategories(datasetId: string | number): Promise<string[]> {
     return getDatasetAvailableCategories.execute(datasetId)
-  }
-
-  getTemplates(collectionIdOrAlias: number | string): Promise<DatasetTemplate[]> {
-    return getDatasetTemplates.execute(collectionIdOrAlias)
   }
 
   link(datasetId: string | number, collectionIdOrAlias: string | number) {