+++ src/stories/dataset/DatasetErrorMockRepository.ts
@@ -11,7 +11,6 @@ import { DatasetVersionSummarySubset } from '@/dataset/domain/models/DatasetVers
 import { DatasetDownloadCount } from '@/dataset/domain/models/DatasetDownloadCount'
 import { CitationFormat, FormattedCitation } from '@/dataset/domain/models/DatasetCitation'
 import { DatasetLicenseUpdateRequest } from '@/dataset/domain/models/DatasetLicenseUpdateRequest'
-import { DatasetTemplate } from '@/dataset/domain/models/DatasetTemplate'
 import { CollectionSummary } from '@/collection/domain/models/CollectionSummary'
 import { DatasetVersionPaginationInfo } from '@/dataset/domain/models/DatasetVersionPaginationInfo'
 
@@ -148,14 +147,6 @@ export class DatasetErrorMockRepository implements DatasetMockRepository {
     _version: string,
     _format: CitationFormat
   ) => {
-    return new Promise((_resolve, reject) => {
-      setTimeout(() => {
-        reject('Error thrown from mock')
-      }, FakerHelper.loadingTimout())
-    })
-  }
-
-  getTemplates(_collectionIdOrAlias: number | string): Promise<DatasetTemplate[]> {
     return new Promise((_resolve, reject) => {
       setTimeout(() => {
         reject('Error thrown from mock')