+++ src/stories/dataset/DatasetMockRepository.ts
@@ -16,8 +16,6 @@ import { DatasetDownloadCount } from '@/dataset/domain/models/DatasetDownloadCou
 import { DatasetDownloadCountMother } from '@tests/component/dataset/domain/models/DatasetDownloadCountMother'
 import { CitationFormat, FormattedCitation } from '@/dataset/domain/models/DatasetCitation'
 import { DatasetLicenseUpdateRequest } from '@/dataset/domain/models/DatasetLicenseUpdateRequest'
-import { DatasetTemplate } from '@/dataset/domain/models/DatasetTemplate'
-import { DatasetTemplateMother } from '@tests/component/dataset/domain/models/DatasetTemplateMother'
 import { CollectionSummary } from '@/collection/domain/models/CollectionSummary'
 import { CollectionSummaryMother } from '@tests/component/collection/domain/models/CollectionSummaryMother'
 import { DatasetVersionPaginationInfo } from '@/dataset/domain/models/DatasetVersionPaginationInfo'
@@ -167,14 +165,6 @@ export class DatasetMockRepository implements DatasetRepository {
           content: 'Formatted citation content',
           contentType: 'text/plain'
         })
-      }, FakerHelper.loadingTimout())
-    })
-  }
-
-  getTemplates(_collectionIdOrAlias: number | string): Promise<DatasetTemplate[]> {
-    return new Promise((resolve) => {
-      setTimeout(() => {
-        resolve(DatasetTemplateMother.createMany(3))
       }, FakerHelper.loadingTimout())
     })
   }