+++ tests/component/sections/shared/dataset-metadata-form/DatasetMetadataForm.spec.tsx
@@ -210,7 +210,6 @@ describe('DatasetMetadataForm', () => {
     datasetRepository.getByPersistentId = cy.stub().resolves(dataset)
     datasetRepository.create = cy.stub().resolves({ persistentId: 'persistentId' })
     datasetRepository.updateMetadata = cy.stub().resolves(undefined)
-    datasetRepository.getTemplates = cy.stub().resolves([])
     metadataBlockInfoRepository.getByCollectionId = cy.stub().resolves(metadataBlocksInfoOnEditMode)
     metadataBlockInfoRepository.getDisplayedOnCreateByCollectionId = cy
       .stub()